Add 2f and 1f prefab door

+++ ProjectEva/Assets/Script/Door/KeyonlyDoor.cs
@@ -8,6 +8,8 @@ public class KeyonlyDoor : MonoBehaviour
 {
     private bool isPlayerNear = false;
     public int hasKeynumber;
+    private string uiPanelTag = "Interactiontag"; 
+    private string customTextTag = "Interactiontext"; 
     public InventoryPresentCharactor inventoryPresentCharactor;
     public GameObject sceneObject;
     public TextMeshProUGUI customText;
@@ -15,7 +17,12 @@ public class KeyonlyDoor : MonoBehaviour
     public string Keyforthisdoor;
     public NavMeshSurface navMeshSurface;
     public int numberofkey;
+    public bool removekey;
     public string custominteractiontext;
+    public void Awake()
+    {
+        FindUIElementsByTag();
+    }
     private void Start()
     {
         inventoryPresentCharactor = FindObjectOfType<InventoryPresentCharactor>();
@@ -67,7 +74,10 @@ public class KeyonlyDoor : MonoBehaviour
     {
         if (isPlayerNear && Input.GetKeyDown(KeyCode.E) && hasKeynumber == numberofkey)
         {
+            if(removekey)
+            {
             inventoryPresentCharactor.DeleteItemCharactorEquipment(Keyforthisdoor);
+            }
             soundManager.PlaySound("Dooropen");
             navMeshSurface.UpdateNavMesh(navMeshSurface.navMeshData);
             Destroy(this.gameObject);
@@ -75,6 +85,22 @@ public class KeyonlyDoor : MonoBehaviour
         if (isPlayerNear && Input.GetKeyDown(KeyCode.E) && hasKeynumber != numberofkey)
         {
             soundManager.PlaySound("Doorlocked");
+        }
+    }
+    private void FindUIElementsByTag()
+    {
+        // Find UI panel by tag
+        GameObject[] sceneObjects = GameObject.FindGameObjectsWithTag(uiPanelTag);
+        if (sceneObjects.Length > 0)
+        {
+            sceneObject = sceneObjects[0]; // Assuming there is only one UI panel with the specified tag
+        }
+
+        // Find custom text by tag
+        GameObject[] customTexts = GameObject.FindGameObjectsWithTag(customTextTag);
+        if (customTexts.Length > 0)
+        {
+            customText = customTexts[0].GetComponent<TextMeshProUGUI>(); // Assuming there is only one custom text with the specified tag
         }
     }
 }
